requirements.txt
appdirs==1.4.0
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
pyparsing==2.1.10
PyYAML==3.12
requests==2.13.0
ruamel.ordereddict==0.4.9
six==1.10.0
typing==3.5.3.0
